+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.8.0]
 ### Added
 - Allow merging schemas by intersection instead of union
 - Use a base schema in `ProductSchema`
@@ -209,7 +211,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2021-05-31
 - Initial release
 
-[Unreleased]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.7.3...HEAD
+[Unreleased]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.7.3...v0.8.0
 [0.7.3]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.7.2...v0.7.3
 [0.7.2]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.7.1...v0.7.2
 [0.7.1]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.7.0...v0.7.1